Refactor and make things build

+++ DIKUArcade/Math/Vec2f.cs
@@ -1,29 +1,29 @@
 ﻿namespace DIKUArcade.Math {
-    public class Vec2i {
-        public int X;
-        public int Y;
+    public class Vec2f {
+        public float X;
+        public float Y;
 
-        public Vec2i(int x, int y) {
+        public Vec2f(float x, float y) {
             X = x;
             Y = y;
         }
 
-        public Vec2i() : this(0, 0) { }
+        public Vec2f() : this(0.0f, 0.0f) { }
 
-        public static Vec2i operator +(Vec2i v1, Vec2i v2) {
-            return new Vec2i(v1.X + v2.X, v1.Y + v2.Y);
+        public static Vec2f operator +(Vec2f v1, Vec2f v2) {
+            return new Vec2f(v1.X + v2.X, v1.Y + v2.Y);
         }
 
-        public static Vec2i operator -(Vec2i v1, Vec2i v2) {
-            return new Vec2i(v1.X - v2.X, v1.Y - v2.Y);
+        public static Vec2f operator -(Vec2f v1, Vec2f v2) {
+            return new Vec2f(v1.X - v2.X, v1.Y - v2.Y);
         }
 
         // pairwise multiplication
-        public static Vec2i operator *(Vec2i v1, Vec2i v2) {
-            return new Vec2i(v1.X * v2.X, v1.Y * v2.Y);
+        public static Vec2f operator *(Vec2f v1, Vec2f v2) {
+            return new Vec2f(v1.X * v2.X, v1.Y * v2.Y);
         }
 
-        public static int Dot(Vec2i v1, Vec2i v2) {
+        public static float Dot(Vec2f v1, Vec2f v2) {
             return v1.X * v2.X + v1.Y * v2.Y;
         }
